src/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.36.0
dj-database-url==2.1.0
Django==4.2.6
django-appconf==1.0.5
django-browser-reload==1.12.0
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-filter==23.4
djangorestframework==3.14.0
idna==3.4
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
rcssmin==1.1.1
requests==2.31.0
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
